+++ src/apps/views/AppsList/AppsList.tsx
@@ -81,7 +81,6 @@ export const AppsList: React.FC<AppsListProps> = ({ params }) => {
 
   const {
     data: appsInProgressData,
-    loading: loadingAppsInProgress,
     refetch: appsInProgressRefetch,
   } = useAppsInstallationsQuery({
     displayLoader: false,
@@ -331,7 +330,6 @@ export const AppsList: React.FC<AppsListProps> = ({ params }) => {
           installedAppsList={installedApps}
           customAppsList={customApps}
           appsInProgressList={appsInProgressData}
-          loadingAppsInProgress={loadingAppsInProgress}
           disabled={loading || customAppsLoading}
           settings={settings}
           onUpdateListSettings={updateListSettings}