+++ src/channels/views/ChannelCreate/ChannelCreate.tsx
@@ -59,19 +59,20 @@ export const ChannelCreateView = ({}) => {
 
   const handleSubmit = async ({
     shippingZonesIdsToAdd,
-    shippingZonesIdsToRemove,
     warehousesIdsToAdd,
-    warehousesIdsToRemove,
     warehousesToDisplay,
-    shippingZonesToDisplay,
     currencyCode,
     allocationStrategy,
-    ...rest
+    name,
+    slug,
+    defaultCountry,
   }: FormData) => {
     const createChannelMutation = createChannel({
       variables: {
         input: {
-          ...rest,
+          defaultCountry,
+          name,
+          slug,
           currencyCode: currencyCode.toUpperCase(),
           addShippingZones: shippingZonesIdsToAdd,
           addWarehouses: warehousesIdsToAdd,